Refactor: Postgres Sync Layer Abstraction
